initial version of pyhttpd
code is mostly based on my http_srv.py of twatscrape project, but
with minor improvements (such as reading files chunk by chunk)
plus support for HTTP/1.1 range requests to receive partially downloaded
files. we only support the subset of range requests necessary to support
resuming downloads, not the additionally specified features[0] that would
allow to specify several ranges or request ranges that end mid-way into
a file.

this is intended as a drop-in replacement for `python -m SimpleHTTPServer`
as shipped with python2, but multi-threaded and with support for download
resuming.

therefore, it currently only supports starting it with no arguments, in
which case it serves contents of current dir on port 8000 on 0.0.0.0
interface, or with a single argument, which is the port number.

apart from that, it's also simple enough to serve custom web applications
as a replacement for e.g. flask with minor modifications.

[0]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Range_requests