rewrite read_request parsing
changes include:
- add BufferedSocket class that features read(), readline() and readuntil()
- improve debug output when client.debugreq = True
- only process POST data when it's in the expected urlencoded form
- request headers are now stored in a dictionary called result['headers'],
  so potentially malignant requests can't unexpectdely overwrite values
  we set ourselves.

the rationale for usage of "BufferedSocket" class is that POST requests
may come in 3 different forms, either text/plain, urlencoded, or multipart,
in the latter case it might include potentially huge files so reading
the entire response into memory isn't smart and might cause the web app
to crash due to OOM on sufficiently big input.

therefore the recipient of the POST request might require to parse incoming
data on its own (for example to write uploaded files in blocks to disk),
however when we eagerly read socket data in blocks we're probably already
past the start of the content data when looking for the initial \r\n\r\n
separator, which would make it impossible to properly postpone the parsing.
the only alternative to buffering the socket data stream is to read the
request byte-by-byte which would cause horrible syscall overhead.