add fileupload demo app
we now handle 2 out of the possible 3 enctypes for POSTed forms
inside pyhttpd itself, those are x-wwwform-urlencoded and text/plain.
the 3rd possible form, multipart/form-data requires special
treatment as it is typically only used for file uploads, and files
are of unknown size so just reading them into memory from the
main app seems a sub-optimal strategy. therefore, we pass the
responsibility to process it to the client app's code, which has
to fetch content-length out of the request dict, and then read
the file data chunk by chunk from its socket object.
a demo-app implementing a working file upload for a single file
is being provided.

use `pyhttpd -a fileupload_demoapp` to test it.

also fix bug parsing non-multipart post-data with separator ':'
instead of '='.