sort script: don't sort unless requested
rather than just providing sort functionality if desired, the
script started to sort the directory listing immediately on load
by column 0 (filename), which causes a noticeable delay when
loading a page containing a sufficiently big amount of items.
additionally, it caused the link to parent directory to not be
displayed on top.